Recreate route in separate file
