--- back-end/routes/badges.js
@@ -1,1 +1,0 @@
-import express from 'express'
